feat(group count): Minimum unit count before issuing command
Modified WaveControl.move_and_attack() so that it no longer accepts a LuaGroup as an argument and no longer sets the command for the group. instead, it returns a table of commands to be used with the compound_command type.
Groups are not issued a command until there is a minimum amount of members. Currently this is a percentage of the total units to be in the group, configurable by modifying config.percent_bugs_created_before_command