: Start of prototyping/added style testing gui
Added a new folder 'data' and a file 'prototypes' that holds the data that we may want to modify, has small and medium biter currently.
Added a gui to test styles. has 'frame', 'dropdown', 'button', and 'progressbar' style types so far. enable in the tools/styles/gui_cheatsheet.lua file, L8 'cheat_gui_enabled = true' and it will be created when your player is